requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==3.0
django-cors-headers==3.2.0
django-extensions==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
